Near-complete rewrite with merged data sets
While I was laying awake last night I realized that the analysis
would be much cleaner if I merged the training and testing data
sets first then passed them through the analysis and summary
pipeline as a whole.  I had been loading them separately.

I've rewritten all of the code to work that way.  It also takes
advantage of the fact that all of the input files for a data
segment have to have corresponding rows.  That means we can
use cbind() instead of merge() and avoid adding subject
and activity IDs until the very end.

This version might be the final one that I submit.  I anticipate
only minor bug fixes from here on out.